+++ packages/luciq_flutter/example/lib/src/screens/my_home_page.dart
@@ -176,6 +176,16 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
+  void _navigateToPrivateViews() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const PrivateViewsPage(),
+        settings: const RouteSettings(name: PrivateViewsPage.screenName),
+      ),
+    );
+  }
+
   final _formUserAttributeKey = GlobalKey<FormState>();
 
   @override
@@ -201,10 +211,12 @@ class _MyHomePageState extends State<MyHomePage> {
           text: 'Bug Reporting',
           symanticLabel: 'open_bug_reporting',
         ),
-        LuciqButton(
-          onPressed: _navigateToCore,
-          text: 'Core',
-          symanticLabel: 'open_core',
+        LuciqPrivateView(
+          child: LuciqButton(
+            onPressed: _navigateToCore,
+            text: 'Core',
+            symanticLabel: 'open_core',
+          ),
         ),
         const SectionTitle('Primary Color'),
         LuciqTextField(
@@ -335,10 +347,12 @@ class _MyHomePageState extends State<MyHomePage> {
           text: 'Show Feature Requests',
           symanticLabel: 'show_feature_requests',
         ),
-        LuciqButton(
-          onPressed: _navigateToCrashes,
-          text: 'Crashes',
-          symanticLabel: 'open_crash_page',
+        LuciqPrivateView(
+          child: LuciqButton(
+            onPressed: _navigateToCrashes,
+            text: 'Crashes',
+            symanticLabel: 'open_crash_page',
+          ),
         ),
         LuciqButton(
           onPressed: _navigateToApm,
@@ -354,6 +368,11 @@ class _MyHomePageState extends State<MyHomePage> {
           onPressed: _navigateToSessionReplay,
           text: 'Session Replay',
           symanticLabel: 'open_session_replay_page',
+        ),
+        LuciqButton(
+          onPressed: _navigateToPrivateViews,
+          text: 'Private Views',
+          symanticLabel: 'open_private_views_page',
         ),
         const SectionTitle('Sessions Replay'),
         LuciqButton(